Adds initial PS script module implementation
This script module is about 80% complete for an initial working
implementation.

Many of the cmdlets were written based on the very same cmdlets that can
be found at https://github.com/Microsoft/SDN in the Kubernetes folder
and at https://github.com/kubernetes-sigs/sig-windows-tools in the
kubeadm folder.

I found that following the documentation, these scripts as defined at
their respective repositories didn't work for one reason or another:

* The ordering was wrong
* Sometimes they used hard-coded "global" variables, other times the
  cmdlets were parameterized
* Hard-coded versions of kubernetes components
* Only work with Windows Server 2019 LTSC2019/1809 (or lower), and not
  with, for example, Windows Server 2019 2004.

So, this script module aims to fix all of these problems. You can use
New-KubernetesClusterNodeConfiguration to create a configuration object,
which is also stored to disk. Or, you can run each and every cmdlet
independently to manually build up your Windows Server-based kubernetes
cluster node. However, at this time, there's a lot of "glue" code in the
New-WindowsKubernetesClusterNode and Join-KubernetesCluster cmdlets. But
it IS possible.

This script module even goes so far as to build custom Docker container
images  for Golang, flannel, and kube-proxy to work with Windows Server
2019 2004.

I'm sure there are many improvements that could be made, but for now, I
just wanted something that WORKS.

The support for building custom kubernetes flannel and kube-proxy images
for Window Server 2019 >1909 has not yet been thouroughly tested.

Since this is the only commit in this repo other than the initial
commit, I will be rebasing this commit--so caveat emptor if cloning this
repository at this time.