WIP: Fixing up a lot of issues -- really close now
This is a WIP commit and will eventually be squashed into the previous
commit when all is said and done. So, beware if you fork/pull from this
repo--I will shamelessly rebase this branch.

This commit fixes A LOT of things.

First, I fixed automating SSH during the node setup process. This
includes uninstalling the OpenSSH Windows Capability and installing the
Powershell/Win32-OpenSSH distribution. This is to get around an issue
with the Windows Server 2019-provided ssh-agent not sending the write
hashes to the remote Linux machine running a newer version of OpenSSH,
causing SSH connections to be refused.

This commit also fixes building the required Docker images. Golang does
not provide a Docker image for Windows Server 2019 2004, which is
required to build the flannel image. And finally, a Windows Server 2019
image compatible with my server version is required for Kube-Proxy, too.

A lot of the orchestration for setting up a Windows Server 2019
kubernetes node has been cleaned up, making it much more streamlined.

One remaining bit of work is to parameterize Join-KubernetesCluster. It
makes a lot of assumptions right now. But, for my purposes at this time,
it's "good enough".

KNOWN ISSUES: The kubeletwin/flannel container image does not stay
running. There's an issue when `wins cli net get` is run. The output is
passed to `ConvertFrom-JSON`, but the output is currently an error. I'm
still trying to determine how to resolve this issue. But this script is
now the closest yet I've come to getting any version of Windows Server
2019 >1809 (or, I should say, 1909) running as a Kubernetes node on
Windows.