WIP: Install wins on the Node's host OS as a service
It turns out that the sig-windows-tools repo's KubeCluster.ps1 and
KubeClusterHelpers.psm1 scripts do not result in Rancher's wins utility
being installed on the Kubernetes node host OS. However, in the same
repo, the PrepareNode.ps1 scripts do accomplish this.

After starting Wins on the host OS and manually running the kube-flannel
pod container, I was able to get a response from `wins cli net get`. So
this is required on the Host OS in order to have a functioning
Kubernetes node.

There were a few other minor cleanups in this commit, too.

I'm hopeful with this commit, that I will be able to successfully have a
functioning Windows Kubernetes Node.