Reorder some things; comment out some things; lots of stuff
OK, still trying to make heads and tails of all the various scripts
which exist "out in the wild" to accomplish setting up a Windows server
as a Kubernetes node. I learned a lot yesterday.

I started with the official instructions at the Kubernetes website,
which tells you to run
http://github.com/kubernetes-sigs/sig-windows-tools/blob/v0.1.3/kubeadm/scripts/PrepareNode.ps1,
and then run a few more commands and "it just works". Except, what I
didn't know is that that's only true if you're running Windows Server
2019 1809 (and now, maybe 1903--and even more maybe 1909--though I've
seen a lot of reports that 1909 doesn't work too well out of the box).

Anyway, that's when I found at the same repo, `KubeCluster.ps1` and
`KubeClusterHelper.psm1`, which had a lot of the same functions over at
the Microsoft/SDN GitHub repo. And so these two scripts became the basis
of this script module.

Fast forward, I learned yesterday that things must have changed. You
don't need to run Flanneld on the node's host OS anymoer. That's covered
by `flannel-overlay.yml` which is applied to the cluster on the control
plane. Same with kube-proxy: you don't need a local service for Kube
Proxy, it runs in a pod in the kube-system namespace by virtue of
applying `kube-proxy.yml`. In a similar manner, because Flannel is not
being installed on the host OS, there's no reason to configure CNI on
the host OS--that's taken care of in the kube-flannel-windows-* pod
container. However, one thing that is missing is Rancher Wins. This is
required to run on the host OS--it allows a container to run commands on
the local Windows host. Note that this is a peculiarity of Windows at
this time.

Anyway, this commit adds support for installing Rancher Wins. It also
changes around the ordering of some things in Join-KubernetesCluster.
But, running these cmdlets still won't get you a working cluter
node--YET. There's still a problem with the kube-proxy-windows pod
container.

So up next are:

1. Resolving the issue with the kube-proxy-windows pod container
    - this will probably result in changing where things are
      "installed", from C:\ProgramData\Kubernetes (where KubeCluster.ps1
      installed things) to C:\k (where PrepareNode.ps1) installs
      things--as I think this is part of the issue. (Actually, the real
      issue may be that I was using c:\k as a working directory while
      running the cmdlets in this script module, but the
      kube-proxy-windows pod container tries to put things in the c:\k
      folder--and so maybe this is causing the container issues. So to
      avoid this--let's just agree to not use c:\k as "scratch space")
2. Remove any unnecessary cmdlets.
    - By my count, a lot of cmdlets have now become superflous. And the
      good news is, they were big ones!
        - Install-ContainerNetworkInterface
	- Install-FlannelD
	- Install-KubeProxy
	- New-NerviceEX
	- New-SerivceWrapper
	- Set-CniConfiguration
	- WaitForNetwork

Also, I've decided that I am NOT going to do any rebasing on this
branch after all. As much as I love a clean commit history, I've learned
too much to rebase it away, and so the history will remain unchanged. So
please feel free to clone this repo if you desire.