Removes a lot of dead code
This commit removes a bulk of the "dead code" that is no longer
required. I'm sure there's stuff I missed. So to determine that, it's
time to refactor the New-WindowsKubernetesClusterNode cmdlet and break
it up into smaller pieces. That should suss out any other code that
needs to be removed. Although, that will also entail adding some
additional cmdlets--don't worry, this is a positive refactoring!