Refactor; fixes remaining issues
This commit largely refactors the New-WindowsKubernetesCluster node
cmdlet by taking all but one or three of the "comma-delimited" sections
of code and puts them in their own set of cmdlets so that this cmdlet
becomes more declarative vs. imperative. It's also a lot shorter and
easier to understand.

Also, this commit contains a very simple change in Install-Kubelet to
almalgamate what can be found in `PrepareNode.ps`' and
`KubeClusterHelper.psm1`. Between both of these scripts, I found the
"magic" combination of configuration settings that, once set
appropriately, allowed my Windows Server 2019 2004 (10.0.19041.508) to
successfully join the Kubernetes cluster as a node in the Ready state.